refactor(bot): remove unnecessary files
