+++ packages/bot/src/config.js
@@ -1,9 +1,9 @@
-'use strict'
+'use strict';
 
-const env = require('./lib/env-helper')
+const env = require('./lib/env-helper');
 
 module.exports = {
-	discordToken: env('TOKEN'),
-	owners: env.array('OWNERS'),
-	backendUrl: env('BACKEND_URL', 'http://localhost:1337'),
-}
+  discordToken: env('TOKEN'),
+  owners: env.array('OWNERS'),
+  backendUrl: env('BACKEND_URL', 'http://localhost:1337'),
+};
