+++ packages/bot/src/listeners/ready.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 const { Listener } = require('@sapphire/framework');
 const { blue, gray, green, magenta, magentaBright, white, yellow } = require('colorette');
@@ -6,55 +6,59 @@ const { blue, gray, green, magenta, magentaBright, white, yellow } = require('co
 const dev = process.env.NODE_ENV !== 'production';
 
 class UserEvent extends Listener {
-	style = dev ? yellow : blue;
+  style = dev ? yellow : blue;
 
-	constructor(context, options = {}) {
-		super(context, {
-			...options,
-			once: true
-		});
-	}
+  constructor(context, options = {}) {
+    super(context, {
+      ...options,
+      once: true,
+    });
+  }
 
-	run() {
-		this.printBanner();
-		this.printStoreDebugInformation();
-	}
+  run() {
+    this.printBanner();
+    this.printStoreDebugInformation();
+  }
 
-	printBanner() {
-		const success = green('+');
+  printBanner() {
+    const success = green('+');
 
-		const llc = dev ? magentaBright : white;
-		const blc = dev ? magenta : blue;
+    const llc = dev ? magentaBright : white;
+    const blc = dev ? magenta : blue;
 
-		const line01 = llc('');
-		const line02 = llc('');
-		const line03 = llc('');
+    const line01 = llc('');
+    const line02 = llc('');
+    const line03 = llc('');
 
-		// Offset Pad
-		const pad = ' '.repeat(7);
+    // Offset Pad
+    const pad = ' '.repeat(7);
 
-		// eslint-disable-next-line no-console
-		console.log(
-			String.raw`
+    // eslint-disable-next-line no-console
+    console.log(
+      String.raw`
 ${line01} ${pad}${blc('1.0.0')}
 ${line02} ${pad}[${success}] Gateway
 ${line03}${dev ? ` ${pad}${blc('<')}${llc('/')}${blc('>')} ${llc('DEVELOPMENT MODE')}` : ''}
 		`.trim()
-		);
-	}
+    );
+  }
 
-	printStoreDebugInformation() {
-		const { client, logger } = this.container;
-		const stores = [...client.stores.values()];
-		const last = stores.pop();
+  printStoreDebugInformation() {
+    const { client, logger } = this.container;
+    const stores = [...client.stores.values()];
+    const last = stores.pop();
 
-		for (const store of stores) logger.info(this.styleStore(store, false));
-		logger.info(this.styleStore(last, true));
-	}
+    for (const store of stores) logger.info(this.styleStore(store, false));
+    logger.info(this.styleStore(last, true));
+  }
 
-	styleStore(store, last) {
-		return gray(`${last ? '└─' : '├─'} Loaded ${this.style(store.size.toString().padEnd(3, ' '))} ${store.name}.`);
-	}
+  styleStore(store, last) {
+    return gray(
+      `${last ? '└─' : '├─'} Loaded ${this.style(store.size.toString().padEnd(3, ' '))} ${
+        store.name
+      }.`
+    );
+  }
 }
 
 module.exports = UserEvent;
