+++ packages/bot/src/preconditions/OwnerOnly.js
@@ -1,12 +1,14 @@
-'use strict'
+'use strict';
 
 const { Precondition } = require('@sapphire/framework');
 const { owners } = require('../config');
 
 class UserPrecondition extends Precondition {
-	async run(message) {
-		return owners.includes(message.author.id) ? this.ok() : this.error({ message: 'This command can only be used by the owners.' });
-	}
+  async run(message) {
+    return owners.includes(message.author.id)
+      ? this.ok()
+      : this.error({ message: 'This command can only be used by the owners.' });
+  }
 }
 
 module.exports = UserPrecondition;
